+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixes issue: https://github.com/supertokens/supertokens-core/issues/258
 - Changes for pagination and count queries: https://github.com/supertokens/supertokens-core/issues/259
 - Add GetThirdPartyUsersByEmail query: https://github.com/supertokens/supertokens-core/issues/277
+- Update user's email via transaction for emailpassword: https://github.com/supertokens/supertokens-core/issues/275
 
 ## [1.8.0] - 2021-04-20
 ### Added